refactor: moves Renderer nodes into its own subfolder
